api/requirements.txt
# Azure Functions
azure-functions>=1.18.0,<2.0.0

# Azure SDK
azure-identity>=1.15.0,<2.0.0
azure-ai-inference>=1.0.0b1,<2.0.0
azure-search-documents>=11.4.0,<12.0.0

# OpenTelemetry
opentelemetry-api>=1.21.0,<2.0.0
opentelemetry-sdk>=1.21.0,<2.0.0
opentelemetry-instrumentation-requests>=0.42b0,<1.0.0
azure-monitor-opentelemetry>=1.2.0,<2.0.0

# JWT validation
PyJWT[crypto]>=2.8.0,<3.0.0
cryptography>=41.0.0,<43.0.0

# HTTP requests
requests>=2.31.0,<3.0.0

# JSON handling
python-dotenv>=1.0.0,<2.0.0
